more work on tests
